renaming emitter file to obj_emitter
